@@ -10,7 +10,6 @@ import java.util.Optional;
 
 import org.jabref.logic.importer.ImportFormatPreferences;
 import org.jabref.logic.importer.fileformat.BibtexImporter;
-import org.jabref.logic.importer.fileformat.BibtexParser;
 import org.jabref.model.entry.BibEntry;
 import org.jabref.model.entry.LinkedFile;
 import org.jabref.model.schema.DublinCoreSchemaCustom;
@@ -29,15 +28,14 @@ import static org.mockito.Mockito.when;
 class XmpUtilReaderTest {
 
     private XmpPreferences xmpPreferences;
-    private BibtexParser parser;
     private BibtexImporter testImporter;
+    private final XmpUtilReader xmpUtilReader = new XmpUtilReader();
 
     /**
      * Create a temporary PDF-file with a single empty page.
      */
     @BeforeEach
     void setUp() {
-        ImportFormatPreferences importFormatPreferences = mock(ImportFormatPreferences.class, Answers.RETURNS_DEEP_STUBS);
         xmpPreferences = mock(XmpPreferences.class);
         // The code assumes privacy filters to be off
         when(xmpPreferences.shouldUseXmpPrivacyFilter()).thenReturn(false);
@@ -53,7 +51,7 @@ class XmpUtilReaderTest {
     @Test
     void testReadArticleDublinCoreReadRawXmp() throws IOException, URISyntaxException {
         Path path = Path.of(XmpUtilShared.class.getResource("article_dublinCore_without_day.pdf").toURI());
-        List<XMPMetadata> meta = XmpUtilReader.readRawXmp(path);
+        List<XMPMetadata> meta = xmpUtilReader.readRawXmp(path);
 
         DublinCoreSchema dcSchema = DublinCoreSchemaCustom.copyDublinCoreSchema(meta.get(0).getDublinCoreSchema());
         DublinCoreExtractor dcExtractor = new DublinCoreExtractor(dcSchema, xmpPreferences, new BibEntry());
@@ -71,7 +69,7 @@ class XmpUtilReaderTest {
     @Test
     void testReadArticleDublinCoreReadXmp() throws IOException, URISyntaxException {
         Path pathPdf = Path.of(XmpUtilShared.class.getResource("article_dublinCore.pdf").toURI());
-        List<BibEntry> entries = XmpUtilReader.readXmp(pathPdf, xmpPreferences);
+        List<BibEntry> entries = xmpUtilReader.readXmp(pathPdf, xmpPreferences);
         Path bibFile = Path.of(XmpUtilShared.class.getResource("article_dublinCore.bib").toURI());
         List<BibEntry> expected = testImporter.importDatabase(bibFile).getDatabase().getEntries();
 
@@ -86,11 +84,11 @@ class XmpUtilReaderTest {
     @Test
     void testReadArticleDublinCoreReadXmpPartialDate() throws IOException, URISyntaxException {
         Path pathPdf = Path.of(XmpUtilShared.class.getResource("article_dublinCore_partial_date.pdf").toURI());
-        List<BibEntry> entries = XmpUtilReader.readXmp(pathPdf, xmpPreferences);
+        List<BibEntry> entries = xmpUtilReader.readXmp(pathPdf, xmpPreferences);
         Path bibFile = Path.of(XmpUtilShared.class.getResource("article_dublinCore_partial_date.bib").toURI());
         List<BibEntry> expected = testImporter.importDatabase(bibFile).getDatabase().getEntries();
 
-        expected.forEach(bibEntry -> bibEntry.setFiles(Arrays.asList(
+        expected.forEach(bibEntry -> bibEntry.setFiles(List.of(
                 new LinkedFile("", pathPdf.toAbsolutePath(), "PDF"))
         ));
 
@@ -102,7 +100,7 @@ class XmpUtilReaderTest {
      */
     @Test
     void testReadEmtpyMetadata() throws IOException, URISyntaxException {
-        List<BibEntry> entries = XmpUtilReader.readXmp(Path.of(XmpUtilShared.class.getResource("empty_metadata.pdf").toURI()), xmpPreferences);
+        List<BibEntry> entries = xmpUtilReader.readXmp(Path.of(XmpUtilShared.class.getResource("empty_metadata.pdf").toURI()), xmpPreferences);
         assertEquals(Collections.emptyList(), entries);
     }
 
@@ -112,12 +110,12 @@ class XmpUtilReaderTest {
     @Test
     void testReadPDMetadata() throws IOException, URISyntaxException {
         Path pathPdf = Path.of(XmpUtilShared.class.getResource("PD_metadata.pdf").toURI());
-        List<BibEntry> entries = XmpUtilReader.readXmp(pathPdf, xmpPreferences);
+        List<BibEntry> entries = xmpUtilReader.readXmp(pathPdf, xmpPreferences);
 
         Path bibFile = Path.of(XmpUtilShared.class.getResource("PD_metadata.bib").toURI());
         List<BibEntry> expected = testImporter.importDatabase(bibFile).getDatabase().getEntries();
 
-        expected.forEach(bibEntry -> bibEntry.setFiles(Arrays.asList(
+        expected.forEach(bibEntry -> bibEntry.setFiles(List.of(
                 new LinkedFile("", pathPdf.toAbsolutePath(), "PDF"))
         ));
 
@@ -129,7 +127,7 @@ class XmpUtilReaderTest {
      */
     @Test
     void testReadNoDescriptionMetadata() throws IOException, URISyntaxException {
-        List<BibEntry> entries = XmpUtilReader.readXmp(Path.of(XmpUtilShared.class.getResource("no_description_metadata.pdf").toURI()), xmpPreferences);
+        List<BibEntry> entries = xmpUtilReader.readXmp(Path.of(XmpUtilShared.class.getResource("no_description_metadata.pdf").toURI()), xmpPreferences);
         assertEquals(Collections.emptyList(), entries);
     }
 }
