@@ -29,7 +29,7 @@ public class XmpUtilReader {
     private static final String START_TAG = "<rdf:Description";
     private static final String END_TAG = "</rdf:Description>";
 
-    private XmpUtilReader() {
+    public XmpUtilReader() {
         // See: https://pdfbox.apache.org/2.0/getting-started.html
         System.setProperty("sun.java2d.cmm", "sun.java2d.cmm.kcms.KcmsServiceProvider"); // To get higher rendering speed on java 8 oder 9 for images
     }
@@ -40,20 +40,10 @@ public class XmpUtilReader {
      * @param path The path to read the XMPMetadata from.
      * @return The XMPMetadata object found in the file
      */
-    public static List<XMPMetadata> readRawXmp(Path path) throws IOException {
-        try (PDDocument document = XmpUtilReader.loadWithAutomaticDecryption(path)) {
-            return XmpUtilReader.getXmpMetadata(document);
+    public List<XMPMetadata> readRawXmp(Path path) throws IOException {
+        try (PDDocument document = loadWithAutomaticDecryption(path)) {
+            return getXmpMetadata(document);
         }
-    }
-
-    /**
-     * Convenience method for readXMP(File).
-     *
-     * @param filename The filename from which to open the file.
-     * @return BibtexEntryies found in the PDF or an empty list
-     */
-    public static List<BibEntry> readXmp(String filename, XmpPreferences xmpPreferences) throws IOException {
-        return XmpUtilReader.readXmp(Path.of(filename), xmpPreferences);
     }
 
     /**
@@ -67,13 +57,13 @@ public class XmpUtilReader {
      * @throws IOException Throws an IOException if the file cannot be read, so the user than remove a lock or cancel
      *                     the operation.
      */
-    public static List<BibEntry> readXmp(Path path, XmpPreferences xmpPreferences)
+    public List<BibEntry> readXmp(Path path, XmpPreferences xmpPreferences)
             throws IOException {
 
         List<BibEntry> result = new LinkedList<>();
 
         try (PDDocument document = loadWithAutomaticDecryption(path)) {
-            List<XMPMetadata> xmpMetaList = XmpUtilReader.getXmpMetadata(document);
+            List<XMPMetadata> xmpMetaList = getXmpMetadata(document);
 
             if (!xmpMetaList.isEmpty()) {
                 // Only support Dublin Core since JabRef 4.2
@@ -82,10 +72,7 @@ public class XmpUtilReader {
                     if (dcSchema != null) {
                         DublinCoreExtractor dcExtractor = new DublinCoreExtractor(dcSchema, xmpPreferences, new BibEntry());
                         Optional<BibEntry> entry = dcExtractor.extractBibtexEntry();
-
-                        if (entry.isPresent()) {
-                            result.add(entry.get());
-                        }
+                        entry.ifPresent(result::add);
                     }
                 }
             }
@@ -111,7 +98,7 @@ public class XmpUtilReader {
      *
      * @return empty List if no metadata has been found, or cannot properly find start or end tag in metadata
      */
-    private static List<XMPMetadata> getXmpMetadata(PDDocument document) {
+    private List<XMPMetadata> getXmpMetadata(PDDocument document) {
         PDDocumentCatalog catalog = document.getDocumentCatalog();
         PDMetadata metaRaw = catalog.getMetadata();
         List<XMPMetadata> metaList = new ArrayList<>();
@@ -154,10 +141,9 @@ public class XmpUtilReader {
      * @param path The path to load.
      * @throws IOException from the underlying @link PDDocument#load(File)
      */
-    public static PDDocument loadWithAutomaticDecryption(Path path) throws IOException {
+    public PDDocument loadWithAutomaticDecryption(Path path) throws IOException {
         // try to load the document
         // also uses an empty string as default password
-        PDDocument doc = Loader.loadPDF(path.toFile());
-        return doc;
+        return Loader.loadPDF(path.toFile());
     }
 }
