@@ -10,50 +10,52 @@ import static org.junit.jupiter.api.Assertions.assertEquals;
 
 class EntryComparatorTest {
 
-    private BibEntry entry1 = new BibEntry();
-    private BibEntry entry2 = new BibEntry();
-
+    @SuppressWarnings("EqualsWithItself")
     @Test
-    void recognizeIdenticObjectsAsEqual() {
-        BibEntry e2 = entry1;
-        assertEquals(0, new EntryComparator(false, false, StandardField.TITLE).compare(entry1, e2));
+    void recognizeIdenticalObjectsAsEqual() {
+        BibEntry entry = new BibEntry();
+        assertEquals(0, new EntryComparator(false, false, StandardField.TITLE).compare(entry, entry));
     }
 
     @Test
-    void compareAuthorFieldBiggerAscending() throws Exception {
-        entry1.setField(StandardField.AUTHOR, "Stephen King");
-        entry2.setField(StandardField.AUTHOR, "Henning Mankell");
-
+    void compareAuthorFieldBiggerAscending() {
+        BibEntry entry1 = new BibEntry()
+                .withField(StandardField.AUTHOR, "Stephen King");
+        BibEntry entry2 = new BibEntry()
+                .withField(StandardField.AUTHOR, "Henning Mankell");
         EntryComparator entryComparator = new EntryComparator(false, false, StandardField.AUTHOR);
 
         assertEquals(-2, entryComparator.compare(entry1, entry2));
     }
 
     @Test
-    void bothEntriesHaveNotSetTheFieldToCompareAscending() throws Exception {
-
-        entry1.setField(StandardField.BOOKTITLE, "Stark - The Dark Half (1989)");
-        entry2.setField(StandardField.COMMENTATOR, "Some Commentator");
+    void bothEntriesHaveNotSetTheFieldToCompareAscending() {
+        BibEntry entry1 = new BibEntry()
+                .withField(StandardField.BOOKTITLE, "Stark - The Dark Half (1989)");
+        BibEntry entry2 = new BibEntry()
+                .withField(StandardField.COMMENTATOR, "Some Commentator");
         EntryComparator entryComparator = new EntryComparator(false, false, StandardField.TITLE);
 
         assertEquals(-1, entryComparator.compare(entry1, entry2));
     }
 
     @Test
-    void secondEntryHasNotSetFieldToCompareAscending() throws Exception {
-
-        entry1.setField(StandardField.TITLE, "Stark - The Dark Half (1989)");
-        entry2.setField(StandardField.COMMENTATOR, "Some Commentator");
+    void secondEntryHasNotSetFieldToCompareAscending() {
+        BibEntry entry1 = new BibEntry()
+                .withField(StandardField.TITLE, "Stark - The Dark Half (1989)");
+        BibEntry entry2 = new BibEntry()
+                .withField(StandardField.COMMENTATOR, "Some Commentator");
         EntryComparator entryComparator = new EntryComparator(false, false, StandardField.TITLE);
 
         assertEquals(-1, entryComparator.compare(entry1, entry2));
     }
 
     @Test
-    void firstEntryHasNotSetFieldToCompareAscending() throws Exception {
-
-        entry1.setField(StandardField.COMMENTATOR, "Some Commentator");
-        entry2.setField(StandardField.TITLE, "Stark - The Dark Half (1989)");
+    void firstEntryHasNotSetFieldToCompareAscending() {
+        BibEntry entry1 = new BibEntry()
+                .withField(StandardField.COMMENTATOR, "Some Commentator");
+        BibEntry entry2 = new BibEntry()
+                .withField(StandardField.TITLE, "Stark - The Dark Half (1989)");
 
         EntryComparator entryComparator = new EntryComparator(false, false, StandardField.TITLE);
 
@@ -61,10 +63,11 @@ class EntryComparatorTest {
     }
 
     @Test
-    void bothEntriesNumericAscending() throws Exception {
-
-        entry1.setField(StandardField.EDITION, "1");
-        entry2.setField(StandardField.EDITION, "3");
+    void bothEntriesNumericAscending() {
+        BibEntry entry1 = new BibEntry()
+                .withField(StandardField.EDITION, "1");
+        BibEntry entry2 = new BibEntry()
+                .withField(StandardField.EDITION, "3");
 
         EntryComparator entryComparator = new EntryComparator(false, false, StandardField.EDITION);
 
@@ -73,29 +76,29 @@ class EntryComparatorTest {
 
     @Test
     void compareObjectsByKeyAscending() {
-        BibEntry e1 = new BibEntry();
-        BibEntry e2 = new BibEntry();
-        e1.setCitationKey("Mayer2019b");
-        e2.setCitationKey("Mayer2019a");
+        BibEntry e1 = new BibEntry()
+                .withCitationKey("Mayer2019b");
+        BibEntry e2 = new BibEntry()
+                .withCitationKey("Mayer2019a");
         assertEquals(1, new EntryComparator(false, false, InternalField.KEY_FIELD).compare(e1, e2));
         assertEquals(-1, new EntryComparator(false, false, InternalField.KEY_FIELD).compare(e2, e1));
     }
 
     @Test
     void compareObjectsByKeyWithNull() {
-        BibEntry e1 = new BibEntry();
+        BibEntry e1 = new BibEntry()
+                .withCitationKey("Mayer2019b");
         BibEntry e2 = new BibEntry();
-        e1.setCitationKey("Mayer2019b");
         assertEquals(-1, new EntryComparator(false, false, InternalField.KEY_FIELD).compare(e1, e2));
         assertEquals(1, new EntryComparator(false, false, InternalField.KEY_FIELD).compare(e2, e1));
     }
 
     @Test
     void compareObjectsByKeyWithBlank() {
-        BibEntry e1 = new BibEntry();
-        BibEntry e2 = new BibEntry();
-        e1.setCitationKey("Mayer2019b");
-        e2.setCitationKey(" ");
+        BibEntry e1 = new BibEntry()
+                .withCitationKey("Mayer2019b");
+        BibEntry e2 = new BibEntry()
+                .withCitationKey(" ");
         assertEquals(-1, new EntryComparator(false, false, InternalField.KEY_FIELD).compare(e1, e2));
         assertEquals(1, new EntryComparator(false, false, InternalField.KEY_FIELD).compare(e2, e1));
     }
