@@ -24,14 +24,12 @@ import static org.mockito.Mockito.when;
 public class XmpExporterTest {
 
     private Exporter exporter;
-    private BibDatabaseContext databaseContext;
+    private final BibDatabaseContext databaseContext = new BibDatabaseContext();
     private final XmpPreferences xmpPreferences = mock(XmpPreferences.class);
 
     @BeforeEach
     public void setUp() {
         exporter = new XmpExporter(xmpPreferences);
-
-        databaseContext = new BibDatabaseContext();
     }
 
     @Test
@@ -44,21 +42,23 @@ public class XmpExporterTest {
 
         exporter.export(databaseContext, file, Collections.singletonList(entry));
         String actual = String.join("\n", Files.readAllLines(file)); // we are using \n to join, so we need it in the expected string as well, \r\n would fail
-        String expected = "  <rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\">\n" +
-                "    <rdf:Description xmlns:dc=\"http://purl.org/dc/elements/1.1/\" rdf:about=\"\">\n" +
-                "      <dc:creator>\n" +
-                "        <rdf:Seq>\n" +
-                "          <rdf:li>Alan Turing</rdf:li>\n" +
-                "        </rdf:Seq>\n" +
-                "      </dc:creator>\n" +
-                "      <dc:format>application/pdf</dc:format>\n" +
-                "      <dc:type>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>Misc</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:type>\n" +
-                "    </rdf:Description>\n" +
-                "  </rdf:RDF>";
+        String expected = """
+                  <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
+                    <rdf:Description xmlns:dc="http://purl.org/dc/elements/1.1/" rdf:about="">
+                      <dc:creator>
+                        <rdf:Seq>
+                          <rdf:li>Alan Turing</rdf:li>
+                        </rdf:Seq>
+                      </dc:creator>
+                      <dc:format>application/pdf</dc:format>
+                      <dc:type>
+                        <rdf:Bag>
+                          <rdf:li>Misc</rdf:li>
+                        </rdf:Bag>
+                      </dc:type>
+                    </rdf:Description>
+                  </rdf:RDF>
+                """.stripTrailing();
         assertEquals(expected, actual);
     }
 
@@ -78,39 +78,41 @@ public class XmpExporterTest {
 
         String actual = String.join("\n", Files.readAllLines(file)); // we are using \n to join, so we need it in the expected string as well, \r\n would fail
 
-        String expected = "  <rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\">\n" +
-                "    <rdf:Description xmlns:dc=\"http://purl.org/dc/elements/1.1/\" rdf:about=\"\">\n" +
-                "      <dc:creator>\n" +
-                "        <rdf:Seq>\n" +
-                "          <rdf:li>Alan Turing</rdf:li>\n" +
-                "        </rdf:Seq>\n" +
-                "      </dc:creator>\n" +
-                "      <dc:format>application/pdf</dc:format>\n" +
-                "      <dc:type>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>Misc</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:type>\n" +
-                "    </rdf:Description>\n" +
-                "    <rdf:Description xmlns:dc=\"http://purl.org/dc/elements/1.1/\" rdf:about=\"\">\n" +
-                "      <dc:creator>\n" +
-                "        <rdf:Seq>\n" +
-                "          <rdf:li>Michael Armbrust</rdf:li>\n" +
-                "        </rdf:Seq>\n" +
-                "      </dc:creator>\n" +
-                "      <dc:relation>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>bibtex/citationkey/Armbrust2010</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:relation>\n" +
-                "      <dc:format>application/pdf</dc:format>\n" +
-                "      <dc:type>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>Misc</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:type>\n" +
-                "    </rdf:Description>\n" +
-                "  </rdf:RDF>";
+        String expected = """
+                    <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
+                      <rdf:Description xmlns:dc="http://purl.org/dc/elements/1.1/" rdf:about="">
+                        <dc:creator>
+                          <rdf:Seq>
+                            <rdf:li>Alan Turing</rdf:li>
+                          </rdf:Seq>
+                        </dc:creator>
+                        <dc:format>application/pdf</dc:format>
+                        <dc:type>
+                          <rdf:Bag>
+                            <rdf:li>Misc</rdf:li>
+                          </rdf:Bag>
+                        </dc:type>
+                      </rdf:Description>
+                      <rdf:Description xmlns:dc="http://purl.org/dc/elements/1.1/" rdf:about="">
+                        <dc:creator>
+                          <rdf:Seq>
+                            <rdf:li>Michael Armbrust</rdf:li>
+                          </rdf:Seq>
+                        </dc:creator>
+                        <dc:relation>
+                          <rdf:Bag>
+                            <rdf:li>bibtex/citationkey/Armbrust2010</rdf:li>
+                          </rdf:Bag>
+                        </dc:relation>
+                        <dc:format>application/pdf</dc:format>
+                        <dc:type>
+                          <rdf:Bag>
+                            <rdf:li>Misc</rdf:li>
+                          </rdf:Bag>
+                        </dc:type>
+                      </rdf:Description>
+                    </rdf:RDF>
+                  """.stripTrailing();
         assertEquals(expected, actual);
     }
 
@@ -122,61 +124,65 @@ public class XmpExporterTest {
 
         BibEntry entryTuring = new BibEntry()
                 .withField(StandardField.AUTHOR, "Alan Turing");
-
         BibEntry entryArmbrust = new BibEntry()
                 .withField(StandardField.AUTHOR, "Michael Armbrust")
                 .withCitationKey("Armbrust2010");
 
         exporter.export(databaseContext, file, List.of(entryTuring, entryArmbrust));
 
+        // Nothing written in given file
         List<String> lines = Files.readAllLines(file);
         assertEquals(Collections.emptyList(), lines);
 
+        // turing contains the turing entry only
         Path fileTuring = Path.of(file.getParent().toString(), entryTuring.getId() + "_null.xmp");
-        String actualTuring = String.join("\n", Files.readAllLines(fileTuring)); // we are using \n to join, so we need it in the expected string as well, \r\n would fail
-
-        String expectedTuring = "  <rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\">\n" +
-                "    <rdf:Description xmlns:dc=\"http://purl.org/dc/elements/1.1/\" rdf:about=\"\">\n" +
-                "      <dc:creator>\n" +
-                "        <rdf:Seq>\n" +
-                "          <rdf:li>Alan Turing</rdf:li>\n" +
-                "        </rdf:Seq>\n" +
-                "      </dc:creator>\n" +
-                "      <dc:format>application/pdf</dc:format>\n" +
-                "      <dc:type>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>Misc</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:type>\n" +
-                "    </rdf:Description>\n" +
-                "  </rdf:RDF>";
-
+        // we are using \n to join, so we need it in the expected string as well, \r\n would fail
+        String actualTuring = String.join("\n", Files.readAllLines(fileTuring));
+        String expectedTuring = """
+                <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
+                  <rdf:Description xmlns:dc="http://purl.org/dc/elements/1.1/" rdf:about="">
+                    <dc:creator>
+                      <rdf:Seq>
+                        <rdf:li>Alan Turing</rdf:li>
+                      </rdf:Seq>
+                    </dc:creator>
+                    <dc:format>application/pdf</dc:format>
+                    <dc:type>
+                      <rdf:Bag>
+                        <rdf:li>Misc</rdf:li>
+                      </rdf:Bag>
+                    </dc:type>
+                  </rdf:Description>
+                </rdf:RDF>
+              """.stripTrailing();
         assertEquals(expectedTuring, actualTuring);
 
+        // armbrust contains the armbrust entry only
         Path fileArmbrust = Path.of(file.getParent().toString(), entryArmbrust.getId() + "_Armbrust2010.xmp");
-        String actualArmbrust = String.join("\n", Files.readAllLines(fileArmbrust)); // we are using \n to join, so we need it in the expected string as well, \r\n would fail
-
-        String expectedArmbrust = "  <rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\">\n" +
-                "    <rdf:Description xmlns:dc=\"http://purl.org/dc/elements/1.1/\" rdf:about=\"\">\n" +
-                "      <dc:creator>\n" +
-                "        <rdf:Seq>\n" +
-                "          <rdf:li>Michael Armbrust</rdf:li>\n" +
-                "        </rdf:Seq>\n" +
-                "      </dc:creator>\n" +
-                "      <dc:relation>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>bibtex/citationkey/Armbrust2010</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:relation>\n" +
-                "      <dc:format>application/pdf</dc:format>\n" +
-                "      <dc:type>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>Misc</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:type>\n" +
-                "    </rdf:Description>\n" +
-                "  </rdf:RDF>";
-
+        // we are using \n to join, so we need it in the expected string as well, \r\n would fail
+        String actualArmbrust = String.join("\n", Files.readAllLines(fileArmbrust));
+        String expectedArmbrust = """
+                  <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
+                    <rdf:Description xmlns:dc="http://purl.org/dc/elements/1.1/" rdf:about="">
+                      <dc:creator>
+                        <rdf:Seq>
+                          <rdf:li>Michael Armbrust</rdf:li>
+                        </rdf:Seq>
+                      </dc:creator>
+                      <dc:relation>
+                        <rdf:Bag>
+                          <rdf:li>bibtex/citationkey/Armbrust2010</rdf:li>
+                        </rdf:Bag>
+                      </dc:relation>
+                      <dc:format>application/pdf</dc:format>
+                      <dc:type>
+                        <rdf:Bag>
+                          <rdf:li>Misc</rdf:li>
+                        </rdf:Bag>
+                      </dc:type>
+                    </rdf:Description>
+                  </rdf:RDF>
+                """.stripTrailing();
         assertEquals(expectedArmbrust, actualArmbrust);
     }
 
@@ -188,21 +194,25 @@ public class XmpExporterTest {
         Path file = testFolder.resolve("ThisIsARandomlyNamedFile");
         Files.createFile(file);
 
-        BibEntry entry = new BibEntry();
-        entry.setField(StandardField.AUTHOR, "Alan Turing");
+        BibEntry entry = new BibEntry()
+                .withField(StandardField.AUTHOR, "Alan Turing");
 
         exporter.export(databaseContext, file, Collections.singletonList(entry));
+
         String actual = String.join("\n", Files.readAllLines(file));
-        String expected = "  <rdf:RDF xmlns:rdf=\"http://www.w3.org/1999/02/22-rdf-syntax-ns#\">\n" +
-                "    <rdf:Description xmlns:dc=\"http://purl.org/dc/elements/1.1/\" rdf:about=\"\">\n" +
-                "      <dc:format>application/pdf</dc:format>\n" +
-                "      <dc:type>\n" +
-                "        <rdf:Bag>\n" +
-                "          <rdf:li>Misc</rdf:li>\n" +
-                "        </rdf:Bag>\n" +
-                "      </dc:type>\n" +
-                "    </rdf:Description>\n" +
-                "  </rdf:RDF>";
+        String expected = """
+                  <rdf:RDF xmlns:rdf="http://www.w3.org/1999/02/22-rdf-syntax-ns#">
+                    <rdf:Description xmlns:dc="http://purl.org/dc/elements/1.1/" rdf:about="">
+                      <dc:format>application/pdf</dc:format>
+                      <dc:type>
+                        <rdf:Bag>
+                          <rdf:li>Misc</rdf:li>
+                        </rdf:Bag>
+                      </dc:type>
+                    </rdf:Description>
+                  </rdf:RDF>
+                """.stripTrailing();
+
         assertEquals(expected, actual);
     }
 }
